pyproject.toml
[project]
name = "music-graph"
version = "0.1.0"
description = "Music data graph from playlist co-occurrence and similarity"
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.23",
    "sqlmodel>=0.0.16",
    "networkx>=3.2",
    "typer>=0.9",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "rapidfuzz>=3.5",
    "musicbrainzngs>=0.7",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "tomli>=2.0",
]

[project.scripts]
music-graph = "music_graph.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/music_graph"]
